rag-qa-system/requirements.txt
# Core Framework
flask==3.0.0
flask-cors==4.0.0
flask-restx==1.3.0
python-dotenv==1.0.0

# LangChain & RAG
langchain==0.2.3
langchain-community==0.2.3
langchain-openai==0.1.8

# Vector Database
chromadb==0.5.0

# Embeddings
openai==1.30.5
sentence-transformers==2.2.2

# Document Processing
pypdf==4.2.0
python-docx==0.8.11
markdown==3.5.2
unstructured==0.11.0
tiktoken==0.7.0

# Utilities
numpy==1.24.3
pandas==2.0.3
pydantic==2.5.0
requests==2.31.0
psutil==5.9.8
redis==5.0.1
sqlalchemy==2.0.23

# Local LLM Support
ollama==0.1.8

# Optional (for better performance)
faiss-cpu==1.7.4
scikit-learn==1.3.0